LayoutTests/imported/w3c/web-platform-tests/css/css-view-transitions/synchronous-callback-skipped-before-run-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: AbortError: Skipping view transition because skipTransition() was called.

Harness Error (FAIL), message = Unhandled rejection: Skipping view transition because skipTransition() was called.

PASS finished promise should be resolved if skipTransition() is invoked before a synchronous updateCallbackDone callback is dispatched
